openinfra.org is really not available (#116)
signed-off-by: yater <vater@bsd.services>
openinfra.org is not available. and openinfra.dev seems to be right.
(btw: all the other are links are fine. :-) )
and feel free to edit this commit (pull request) especially for "violating" "ow rules" reasons.
sorry for using this repository "to play around with github and dco (signed-off-by foo) for the first time".