First draft for openSSL-3 security advisory. (#440)
* First draft for openSSL-3 security advisory.

* it is OpenSSL, not openSSL.

Reference: openssl.org

* that's -> that is, nicer written english

* be consistent about redeploying

* Add list of k8s services and their OpenSSL status.

* Fix table format.

* Add FKR as author. Reword note on Ubu 22.04.

* Remove section on our preparedness.

Well, we worked out on that before and it is not strictly required here,
as @fkr correctly commented. Shorten the reminder on fixing laptops.

* More precise language on fixing applications.

Also mention openssl version command.

* Add short description, links and image.

* Remove question marks from golang services.

I've looked into the source code and found them to be pure golang
builds. (This might also be true for e.g. OCCM, but I saw openSSL
in the container image there nevertheless.)

* Adjust pub date to 18:30, add Ubu adv link.

* Pub date 18:45 CET. Add _de symlink.

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Signed-off-by: Felix Kronlage-Dammers <fkr@hazardous.org>
Co-authored-by: Felix Kronlage-Dammers <fkr@hazardous.org>