Fix/contribute link (#778)
* Link Community section in Docs pages.

The Community section there is the best description of how we interact
and work, so link it rather than having increasingly stale information
here.

Note: I did not yet remove any content here, not having double-checked
that everything is covered on the docs pages. Eventually, we should
remove these pages from scs.community and link directly to the relevant
pages on docs.scs.community.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Minimize contribute page.

Next step will be to remove it and instead link the docs communty
page directly from the nav menu. To do so, move the section on
international community and the invitation to contact us and
the map to the use page.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Remove contribute pages (but leave redirect).

The map of contributors has been moved to the use page.
The old /contibute page is still there, redirecting to docs.../community
The menubar should link to docs directly.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

* Rely on redirect of /contribute as menubar needs it.

Also a few minor corrections on our use page.

Signed-off-by: Kurt Garloff <kurt@garloff.de>

---------

Signed-off-by: Kurt Garloff <kurt@garloff.de>
Co-authored-by: Max Wolfs <mail@maxwolfs.com>