Addressing #1095, step 1: Link symbols (conditional)
Some Conference contribution entries have a link, some not.
(Most should!)
The ones that do support clicking on the item to jump to it.
The ones that don't have an empty link, resulting in the browser to
reload the current page. Mouse over is the only way to fint out for
users that I'm aware of.

Change two thing:
- Avoid empty links.
- Add a link symbol if we do have a link.

Signed-off-by: Kurt Garloff <kurt@garloff.de>